// client/src/app/services/api/danbooru/models.ts
export interface DanbooruPost {
  id: number;
  created_at: string;
  uploader_id: number;
  score: number;
  source: string;
  md5: string;
  last_comment_bumped_at: string | null;
  rating: 'g' | 's' | 'q' | 'e';
  image_width: number;
  image_height: number;
  tag_string: string;
  tag_string_general: string;
  tag_string_character: string;
  tag_string_copyright: string;
  tag_string_artist: string;
  tag_string_meta: string;
  file_url: string;
  large_file_url: string;
  preview_file_url: string;
  file_ext: string;
  file_size: number;
  is_deleted: boolean;
  is_banned: boolean;
  fav_count: number;
  pool_string: string;
  up_score: number;
  down_score: number;
  updated_at: string;
  pixiv_id: number | null;
  [key: string]: unknown;
}